refactor: make infrastructure layer not depend on application layer
