+++ src/main/java/com/bmi/bookmarkitapi/bookmark/infrastructure/BookmarkClassifiedEventConsumer.java
@@ -1,23 +1,22 @@
 package com.bmi.bookmarkitapi.bookmark.infrastructure;
 
-import com.bmi.bookmarkitapi.bookmark.application.BookmarkModificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.StringTokenizer;
 
 @Slf4j
 @RequiredArgsConstructor
-@Service
+@Component
 public class BookmarkClassifiedEventConsumer {
 
     private final BookmarkModificationService bookmarkModificationService;
 
     @RabbitListener(queues = "q.classification.classified")
     public void listen(String message) {
-        StringTokenizer stringTokenizer = new StringTokenizer(message, "|");
+        StringTokenizer stringTokenizer = new StringTokenizer(message, "|||");
 
         Long bookmarkId = Long.parseLong(stringTokenizer.nextToken());
         String content = stringTokenizer.nextToken();
@@ -29,7 +28,6 @@ public class BookmarkClassifiedEventConsumer {
         log.info("mainCategory: " + mainCategory);
         log.info("subCategory: " + subCategory);
 
-        bookmarkModificationService.setContentAndCategory(bookmarkId, content, mainCategory, subCategory);
-        bookmarkModificationService.complete(bookmarkId);
+        bookmarkModificationService.complete(bookmarkId, content, mainCategory, subCategory);
     }
 }
